Test and add HydrationRepo.getMostHydratedUser()
