+++ src/ActivityRepo.js
@@ -107,14 +107,6 @@ class ActivityRepo {
     const userInfo = data.find(user => user.id === sortedUsers[0].userID)
     return userInfo.name
   }
-
-  getBestSleptUser(date, userData) {
-    const entries = this.findDailySleepForAll(date)
-    const sortedSleepers = entries.sort((entryA, entryB) => entryB.sleepQuality - entryA.sleepQuality)
-    const bestSleeper = sortedSleepers[0]
-    const bestSleeperName = userData.find(user => user.id === bestSleeper.userID)
-    return bestSleeperName.name
-  }
 }
 
 if (typeof module !== 'undefined') {
